Fixed - multiple deaths - level increased on lives

+++ src/game/domain/level/behavior/PlayerDeathBehavior.kt
@@ -1,6 +1,7 @@
 package game.domain.level.behavior
 
 import game.JBomb
+import game.domain.events.game.DefeatGameEvent
 import game.domain.world.domain.entity.actors.abstracts.entity_interactable.EntityInteractable
 import game.presentation.ui.pages.game_over.GameOverPanel
 import java.awt.event.ActionEvent
@@ -8,11 +9,13 @@ import javax.swing.Timer
 
 class PlayerDeathBehavior : GameBehavior() {
     override fun hostBehavior(): () -> Unit = {
-        if (JBomb.match.currentLevel.gameHandler.canGameBeEnded()) {
+        if (JBomb.match.currentLevel.gameHandler.canGameBeEnded() && !JBomb.isGameEnded) {
             val t = Timer(EntityInteractable.SHOW_DEATH_PAGE_DELAY_MS.toInt()) { _: ActionEvent? ->
-                JBomb.match.currentLevel.endLevel()
-                JBomb.destroyLevel(true)
-                JBomb.showActivity(GameOverPanel::class.java)
+                if (!JBomb.isGameEnded) {
+                    DefeatGameEvent().invoke(null)
+                    JBomb.destroyLevel(true)
+                    JBomb.showActivity(GameOverPanel::class.java)
+                }
             }
             t.isRepeats = false
             t.start()
